requirements.txt
anyascii==0.3.2
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.4
django-bakery==0.13.2
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
fs==2.4.16
html5lib==1.1
idna==3.7
jmespath==1.0.1
l18n==2021.3
laces==0.1.1
openpyxl==3.1.2
pillow==10.3.0
pillow_heif==0.16.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
telepath==0.3.1
tzdata==2024.1
urllib3==2.2.1
wagtail==6.0.2
webencodings==0.5.1
Willow==1.8.0
